ref(crons): Split out MonitorEnvironmentLabel
